docs: Give categories weight for better IA
The docs navigation is currently jumbled, without a consistent
order for documentation. This PR gives weight to the content -
the highest for the Overview, so it sticks to the top, then a descending
order for the spin-operator, shim, class-manager, and spin-plugin based
partially on the order that we expect folks to discover them in.

Signed-off-by: Danielle Lancashire <dani@builds.terrible.systems>